Extract media detail query service
